# swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: API V1
  version: v1
paths:
...
  "/reports/api/v1/reports/{report_id}/item_accounts/create_batch":
    post:
      summary: create item_accounts
      tags:
      - ItemAccount
      operationId: create_batch
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '201':
          description: successful
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/item_account"
        required: true
...
servers:
- url: https://{defaultHost}
  variables:
    defaultHost:
      default: www.example.com
components:
  schemas:
...
    item_account:
      type: object
      properties:
        id:
          type: string
        report_service_id:
          type: integer
        report_id:
          type: string
        parent_item_id:
          type: string
        child_item_id:
          type: string
      required:
      - parent_item_id
      - child_item_id